feat: Add Dockerfile and Makefile for aws-kubectl
This commit adds a Dockerfile and Makefile to the aws-kubectl directory. The Dockerfile installs curl, kubectl, and awscli in the Python Alpine image. The Makefile includes a "push" target that builds the docker image and pushes it to kmltrk/aws-kubectl:latest repository.

Also, update manifest.registry.yaml with new image reference

In addition, the manifest.registry.yaml file is updated to use the latest image from kmltrk/aws-kubectl repository instead of odaniait/aws-kubectl:latest. The AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY fields are rearranged for consistency.