Add ECR registry job to deploy Docker credentials
This commit adds a new YAML file `job.registry.yaml` that defines a Kubernetes Job resource. The job is responsible for retrieving Docker credentials for the AWS ECR Registry and updating the corresponding secret in the specified namespace. It uses an image `kmltrk/aws-kubectl:0.0.1` and sets environment variables using AWS CLI commands. The updated secret includes the ECR token retrieved from AWS, and it is created or updated in the specified namespace.

The commit also includes other minor changes related to formatting and comments.